+++ Sources/CPU/Instruction/Arithmethic & Logic/SRA.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct SRA: Instruction {
-    let cycles: Int = 1
+    let cycles: UInt16 = 1
     let target: Target
     
     func execute(with cpu: inout CPU) throws {
